python/requirements.txt
fastapi
uvicorn
pydantic
numpy
pillow
torch
torchvision
onnxruntime
python-multipart